requirements.txt
openai==1.107.1
PyGithub==1.52.0
semgrep==1.35.0
python-dotenv==1.0.0
